@@ -9,24 +9,42 @@ exclude = ["local_ignore"]
 module = "cloudpickle.*,fiona.*,geopandas.*,matplotlib.*,osgeo.*,osgeo_utils.*,pygeos.*,pyogrio.*,psutil.*,setuptools.*,shapely.*,topojson.*"
 ignore_missing_imports = true
 
+[tool.pytest.ini_options]
+log_level = "DEBUG"
+# Write all logging and warnings immediately. Useful for debugging.
+# log_cli = true
+# addopts = "-p no:warnings"
+
+# Some filters that apply to warnings that can be ignored for any test.
+filterwarnings = [
+    "ignore: The output driver does not natively support DateTime type",
+    "ignore: Field .* create as date field, though DateTime requested",
+    "ignore: Geometry is in a geographic CRS",
+    "ignore: Normalized/laundered field name",
+    "ignore: Non-conformant content for record 0 in column",
+    "ignore: Non-conformant content for record 1 in column",
+]
+
 [tool.ruff]
 line-length = 88
+target-version = "py39"
+extend-exclude = ["docs/*", "local_ignore/*"]
+
+[tool.ruff.lint]
 select = [
     # pyflakes
     "F",
     # pycodestyle
     "E",
     "W",
-    # flake8-2020
-    "YTT",
+    # pyupgrade
+    "UP",
     # flake8-bugbear
     "B",
-    # flake8-quotes
-    "Q",
     # flake8-debugger
     "T10",
-    # flake8-gettext
-    "INT",
+    # flake8-simplify
+    # "SIM",
     # pylint
     "PLC",
     "PLE",
