@@ -1,6 +1,4 @@
-"""
-Setup file to package geofileops.
-"""
+"""Setup file to package geofileops."""
 
 import setuptools
 
@@ -20,12 +18,14 @@ setuptools.setup(
     long_description_content_type="text/markdown",
     url="https://github.com/geofileops/geofileops",
     include_package_data=True,
-    packages=setuptools.find_packages(),
+    packages=setuptools.find_packages(
+        exclude=("tests", "benchmark", "benchmark.*"),
+        include=("geofileops", "geofileops.*"),
+    ),
     install_requires=[
         "cloudpickle",
-        "fiona",
-        "gdal>=3.6",
-        "geopandas>=0.12,<1",
+        "gdal>=3.6,<3.10",
+        "geopandas>=0.12,<1.1",
         "numpy",
         "packaging",
         "pandas",
@@ -37,8 +37,12 @@ setuptools.setup(
     ],
     extras_require={"full": ["simplification"]},
     classifiers=[
-        "Programming Language :: Python :: 3",
+        "Intended Audience :: Science/Research",
+        "License :: OSI Approved :: BSD License",
         "Operating System :: OS Independent",
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3 :: Only",
+        "Topic :: Scientific/Engineering :: GIS",
     ],
-    python_requires=">=3.8",
+    python_requires=">=3.9",
 )
