@@ -1,49 +1,52 @@
-"""
-Module containing the implementation of Geofile operations using GeoPandas.
-"""
+"""Module containing the implementation of Geofile operations using GeoPandas."""
 
-from concurrent import futures
 import copy
-from datetime import datetime
 import enum
 import json
 import logging
 import logging.config
 import math
 import multiprocessing
-from pathlib import Path
 import pickle
 import re
 import shutil
 import time
-from typing import Any, Callable, Dict, Iterable, List, Optional, Set, Tuple, Union
 import warnings
+from collections.abc import Iterable
+from concurrent import futures
+from datetime import datetime
+from pathlib import Path
+from typing import Any, Callable, Optional, Union
 
 import cloudpickle
 import geopandas as gpd
 import numpy as np
 import pandas as pd
-import pygeoops
 import psutil
-
-from pygeoops import GeometryType, PrimitiveType
+import pygeoops
 import shapely
 import shapely.geometry as sh_geom
+from pygeoops import GeometryType, PrimitiveType
 
 import geofileops as gfo
 from geofileops import fileops
-from geofileops.helpers._configoptions_helper import ConfigOptions
-
+from geofileops._compat import GEOPANDAS_GTE_10, PANDAS_GTE_22
 from geofileops.helpers import _parameter_helper
-from geofileops.util import _general_util
-from geofileops.util import _geoops_sql
-from geofileops.util import _geoseries_util
-from geofileops.util import _io_util
-from geofileops.util import _ogr_util
-from geofileops.util import _processing_util
+from geofileops.helpers._configoptions_helper import ConfigOptions
+from geofileops.util import (
+    _general_util,
+    _geoops_sql,
+    _geoseries_util,
+    _io_util,
+    _ogr_util,
+    _processing_util,
+)
 from geofileops.util._geofileinfo import GeofileInfo
-from geofileops.util._geometry_util import SimplifyAlgorithm
-from geofileops.util._geometry_util import BufferEndCapStyle, BufferJoinStyle
+from geofileops.util._geometry_util import (
+    BufferEndCapStyle,
+    BufferJoinStyle,
+    SimplifyAlgorithm,
+)
 
 # Don't show this geopandas warning...
 warnings.filterwarnings("ignore", "GeoSeries.isna", UserWarning)
@@ -52,8 +55,7 @@ logger = logging.getLogger(__name__)
 
 
 class ParallelizationConfig:
-    """
-    Heuristics for geopandas based geo operations.
+    """Heuristics for geopandas based geo operations.
 
     Heuristics meant to be able to optimize the parallelisation parameters for
     geopandas based geo operation.
@@ -69,8 +71,7 @@ class ParallelizationConfig:
         bytes_usable: Optional[int] = None,
         cpu_count: int = -1,
     ):
-        """
-        Heuristics for geopandas based geo operations.
+        """Heuristics for geopandas based geo operations.
 
         Heuristics meant to be able to optimize the parallelisation parameters for
         geopandas based geo operation.
@@ -126,9 +127,8 @@ def _determine_nb_batches(
     nb_parallel: int = -1,
     batchsize: int = -1,
     parallelization_config: Optional[ParallelizationConfig] = None,
-) -> Tuple[int, int]:
-    """
-    Determines recommended parallelization params.
+) -> tuple[int, int]:
+    """Determines recommended parallelization params.
 
     Args:
         nb_rows_total (int): The total number of rows that will be processed
@@ -240,7 +240,7 @@ class ProcessingParams:
         self,
         nb_rows_to_process: int,
         nb_parallel: int,
-        batches: List[str],
+        batches: list[str],
         batchsize: int,
     ):
         self.nb_rows_to_process = nb_rows_to_process
@@ -272,7 +272,7 @@ def _prepare_processing_params(
     )
 
     # Prepare batches to process
-    batches: List[str] = []
+    batches: list[str] = []
     if nb_batches == 1:
         # If only one batch, no filtering is needed
         batches.append("")
@@ -373,7 +373,7 @@ def apply(
     only_geom_input: bool = True,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     force_output_geometrytype: Union[GeometryType, str, None] = None,
     gridsize: float = 0.0,
@@ -424,7 +424,7 @@ def buffer(
     single_sided: bool = False,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
@@ -476,7 +476,7 @@ def convexhull(
     output_path: Path,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
@@ -486,7 +486,7 @@ def convexhull(
     force: bool = False,
 ):
     # Init
-    operation_params: Dict[str, Any] = {}
+    operation_params: dict[str, Any] = {}
 
     # Go!
     return _apply_geooperation_to_layer(
@@ -513,7 +513,7 @@ def makevalid(
     output_path: Path,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     force_output_geometrytype: Union[str, None, GeometryType] = None,
     gridsize: float = 0.0,
@@ -566,7 +566,7 @@ def simplify(
     lookahead: int = 8,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
@@ -608,7 +608,7 @@ def _apply_geooperation_to_layer(
     operation: GeoOperation,
     operation_params: dict,
     input_layer: Optional[str],  # = None
-    columns: Optional[List[str]],  # = None
+    columns: Optional[list[str]],  # = None
     output_layer: Optional[str],  # = None
     explodecollections: bool,  # = False
     force_output_geometrytype: Union[GeometryType, str, None],  # = None
@@ -620,8 +620,7 @@ def _apply_geooperation_to_layer(
     force: bool,  # = False
     parallelization_config: Optional[ParallelizationConfig] = None,
 ):
-    """
-    Applies a geo operation on a layer.
+    """Applies a geo operation on a layer.
 
     The operation to apply can be one of the the following:
       - BUFFER: apply a buffer. Operation parameters:
@@ -711,18 +710,17 @@ def _apply_geooperation_to_layer(
         raise ValueError(f"{operation_name}: output_path must not equal input_path")
     if input_layer is None:
         input_layer = gfo.get_only_layer(input_path)
-    if output_path.exists():
-        if force is False:
-            logger.info(f"Stop, output exists already {output_path}")
-            return
-        else:
-            gfo.remove(output_path)
+    if _io_util.output_exists(path=output_path, remove_if_exists=force):
+        return
     if input_layer is None:
         input_layer = gfo.get_only_layer(input_path)
     if output_layer is None:
         output_layer = gfo.get_default_layer(output_path)
     if isinstance(force_output_geometrytype, GeometryType):
         force_output_geometrytype = force_output_geometrytype.name
+    if isinstance(columns, str):
+        # If a string is passed, convert to list
+        columns = [columns]
 
     # Check if we want to preserve the fid in the output
     preserve_fid = False
@@ -771,7 +769,7 @@ def _apply_geooperation_to_layer(
             # Prepare output filename
             tmp_output_path = tmp_dir / output_path.name
 
-            batches: Dict[int, dict] = {}
+            batches: dict[int, dict] = {}
             future_to_batch_id = {}
 
             for batch_id, batch_filter in enumerate(processing_params.batches):
@@ -897,7 +895,7 @@ def _apply_geooperation(
     operation_params: dict,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     where=None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
@@ -906,6 +904,8 @@ def _apply_geooperation(
     force: bool = False,
 ) -> str:
     # Init
+    if not output_path.parent.exists():
+        raise ValueError(f"Output directory does not exist: {output_path.parent}")
     if output_path.exists():
         if force is False:
             message = f"Stop, output exists already {output_path}"
@@ -1023,22 +1023,60 @@ def dissolve(
     batchsize: int = -1,
     force: bool = False,
     operation_prefix: str = "",
-) -> dict:
-    """
-    Function that applies a dissolve.
+):
+    """Function that applies a dissolve.
 
-    More detailed documentation in module geoops!
+    End user documentation can be found in module geoops!
 
     Remark: keep_empty_geoms is not implemented because this is not so easy because
     (for polygon dissolve) the batches are location based, and null/empty geometries
     don't have a location. It could be implemented, but as long as nobody needs it...
+
+    The attribute data aggregation logic is a bit more complex to be able to process
+    per tile and in multiple passed for large datasets:
+      - Note that a geometry that lies on the edge of 2 (or more) tiles will be split up
+        on the tile boundary(ies) and each part will be further treated in the
+        respective tile.
+      - To be able to correctly perform attribute aggregations, they can only be
+        determined after all tiles and passes have been finished, as the information
+        from multiple tiles over multiple passes might have to be combined.
+      - Hence, all needed data (columns and values) is stored in intermediate/temporary
+        results so it can be all combined at the end.
+      - In practice, during the first calculation pass, all relevant columns and values
+        as well as the original fid of the geometries are serialized as a JSON string
+        for each input geometry. When a geometry is dissolved with another geometry in
+        this pass, their json strings are concatenated to a list. This way, all data is
+        retained. An example of a JSON string list for 2 dissolved geometries:
+            [{"fid_orig": 1, "area": 10.0}, {"fid_orig": 2, "area": 5.0}]
+      - When geometries are merged in a following dissolve pass, the lists of JSON
+        strings will be concatenated so all data is always retained. If a geometry was
+        on the border of 2 tiles, this can result in multiple identical JSON strings. In
+        the following example, fid_orig 1 was on the border of 2 tiles and was dissolved
+        again in a following pass, leading to the following JSON string list:
+            [
+                {"fid_orig": 1, "area": 10.0},
+                {"fid_orig": 1, "area": 10.0},
+                {"fid_orig": 2, "area": 5.0},
+            ]
+      - When all passes are done, meaning everything is glued together and all attribute
+        JSON strings are combined in one big list for each final geometry, the attribute
+        aggregations can be performed.
+      - When an original geometry was on the boundary of 2 (or more) tiles in the first
+        pass, like in the example above, the aggregation has to ignore the resulting
+        duplicate atttribute JSON strings. Otherwise a e.g. "SUM" aggregate will
+        double-count values.
+      - The `fid_orig` with the original `fid` from the source file is included for this
+        reason. The `fid_orig` and all attributes of such split geometries will be the
+        same. The `fid_orig` of other geometries that were actually dissolved will
+        always be different. Hence, a simple "distinct" on the JSON strings will result
+        in the correct list of JSON strings that should be used to base the agregations
+        on. The only caveat is that the order of the columns in the JSON strings always
+        needs to be the same.
     """
     # Init and validate input parameters
     # ----------------------------------
-    start_time = datetime.now()
     operation_name = f"{operation_prefix}dissolve"
     logger = logging.getLogger(f"geofileops.{operation_name}")
-    result_info = {}
 
     # Check input parameters
     if groupby_columns is not None and len(list(groupby_columns)) == 0:
@@ -1105,15 +1143,11 @@ def dissolve(
                 )
 
     # Now input parameters are checked, check if we need to calculate anyway
-    if output_path.exists():
-        if force is False:
-            result_info[
-                "message"
-            ] = f"Stop, output exists already {output_path} and force is false"
-            logger.info(result_info["message"])
-            return result_info
-        else:
-            gfo.remove(output_path)
+    if _io_util.output_exists(path=output_path, remove_if_exists=force):
+        return
+
+    # Check what we need to do in an error occurs
+    on_data_error = ConfigOptions.on_data_error
 
     # Now start dissolving
     # --------------------
@@ -1305,6 +1339,7 @@ def dissolve(
                     gridsize=gridsize,
                     keep_empty_geoms=False,
                     nb_parallel=nb_parallel,
+                    on_data_error=on_data_error,
                 )
                 logger.info(f"Pass {pass_id} ready, took {datetime.now()-pass_start}")
 
@@ -1559,11 +1594,6 @@ def dissolve(
             f"Unsupported input geometrytype: {input_layerinfo.geometrytype}"
         )
 
-    # Return result info
-    result_info["message"] = f"Ready, took {datetime.now()-start_time}"
-    logger.info(result_info["message"])
-    return result_info
-
 
 def _dissolve_polygons_pass(
     input_path: Path,
@@ -1578,6 +1608,7 @@ def _dissolve_polygons_pass(
     gridsize: float,
     keep_empty_geoms: bool,
     nb_parallel: int,
+    on_data_error: str = "raise",
 ):
     start_time = datetime.now()
 
@@ -1597,7 +1628,7 @@ def _dissolve_polygons_pass(
         # Prepare output filename
         tempdir = output_onborder_path.parent
 
-        batches: Dict[int, dict] = {}
+        batches: dict[int, dict] = {}
         nb_batches = len(tiles_gdf)
         nb_batches_done = 0
         future_to_batch_id = {}
@@ -1612,14 +1643,14 @@ def _dissolve_polygons_pass(
             suffix = output_notonborder_path.suffix
             name = f"{output_notonborder_path.stem}_{batch_id}{suffix}"
             output_notonborder_tmp_partial_path = tempdir / name
-            batches[batch_id][
-                "output_notonborder_tmp_partial_path"
-            ] = output_notonborder_tmp_partial_path
+            batches[batch_id]["output_notonborder_tmp_partial_path"] = (
+                output_notonborder_tmp_partial_path
+            )
             name = f"{output_onborder_path.stem}_{batch_id}{suffix}"
             output_onborder_tmp_partial_path = tempdir / name
-            batches[batch_id][
-                "output_onborder_tmp_partial_path"
-            ] = output_onborder_tmp_partial_path
+            batches[batch_id]["output_onborder_tmp_partial_path"] = (
+                output_onborder_tmp_partial_path
+            )
 
             # Get tile_id if present
             tile_id = tile_row.tile_id if "tile_id" in tile_row._fields else None
@@ -1639,6 +1670,7 @@ def _dissolve_polygons_pass(
                 tile_id=tile_id,
                 gridsize=gridsize,
                 keep_empty_geoms=keep_empty_geoms,
+                on_data_error=on_data_error,
             )
             future_to_batch_id[future] = batch_id
 
@@ -1723,13 +1755,14 @@ def _dissolve_polygons(
     input_geometrytype: GeometryType,
     input_layer: Optional[str],
     output_layer: Optional[str],
-    bbox: Tuple[float, float, float, float],
+    bbox: tuple[float, float, float, float],
     tile_id: Optional[int],
     gridsize: float,
     keep_empty_geoms: bool,
+    on_data_error: str = "raise",
 ) -> dict:
     # Init
-    perfinfo: Dict[str, float] = {}
+    perfinfo: dict[str, float] = {}
     start_time = datetime.now()
     return_info = {
         "input_path": input_path,
@@ -1749,7 +1782,7 @@ def _dissolve_polygons(
     agg_columns_needed = None
     while True:
         try:
-            columns_to_read: Set[str] = set()
+            columns_to_read: set[str] = set()
             info = gfo.get_layerinfo(input_path, input_layer)
             if groupby_columns is not None:
                 columns_to_read.update(groupby_columns)
@@ -1761,15 +1794,20 @@ def _dissolve_polygons(
                     # are already in the json column
                     columns_to_read.add("__DISSOLVE_TOJSON")
                 else:
-                    # The first pass, so read all relevant columns to code
-                    # them in json
+                    # The first pass, so read all relevant columns to code them in json
                     if "json" in agg_columns:
-                        agg_columns_needed = agg_columns["json"]
+                        agg_columns_needed = list(agg_columns["json"])
                     elif "columns" in agg_columns:
                         agg_columns_needed = [
                             agg_column["column"]
                             for agg_column in agg_columns["columns"]
                         ]
+
+                        # Avoid reading/saving needed columns multiple times.
+                        # The order of the columns should always be the same in the json
+                        # to be able to filter distinct rows efficiently, so sort them,
+                        # as a set gives a different order from run to run.
+                        agg_columns_needed = sorted(set(agg_columns_needed))
                     if agg_columns_needed is not None:
                         columns_to_read.update(agg_columns_needed)
 
@@ -1781,10 +1819,17 @@ def _dissolve_polygons(
                 fid_as_index=fid_as_index,
             )
 
-            if agg_columns is not None:
-                input_gdf["fid_orig"] = input_gdf.index
-                if agg_columns_needed is not None:
-                    agg_columns_needed.append("fid_orig")
+            if agg_columns is not None and agg_columns_needed is not None:
+                # The fid should be added as well, but make name unique
+                fid_orig_column = "fid_orig"
+                for idx in range(99999):
+                    if idx != 0:
+                        fid_orig_column = f"fid_orig{idx}"
+                    if fid_orig_column not in agg_columns_needed:
+                        break
+
+                input_gdf[fid_orig_column] = input_gdf.index
+                agg_columns_needed.insert(0, fid_orig_column)
 
             break
         except Exception as ex:
@@ -1811,17 +1856,37 @@ def _dissolve_polygons(
     aggfunc: Union[str, dict, None] = None
     if agg_columns is not None:
         if "__DISSOLVE_TOJSON" not in input_gdf.columns:
-            # First pass -> put relevant columns in json field
+            # First pass -> put relevant columns in json field.
             aggfunc = {"to_json": agg_columns_needed}
         else:
             # Columns already coded in a json column, so merge json lists
             aggfunc = "merge_json_lists"
     else:
         aggfunc = "first"
+
     start_dissolve = datetime.now()
-    diss_gdf = _dissolve(
-        df=input_gdf, by=groupby_columns, aggfunc=aggfunc, as_index=False, dropna=False
-    )
+    try:
+        diss_gdf = _dissolve(
+            df=input_gdf,
+            by=groupby_columns,
+            aggfunc=aggfunc,
+            as_index=False,
+            dropna=False,
+        )
+    except Exception as ex:
+        # If a GEOS exception occurs, it is probably due to invalid geometries.
+        # Try to fix them and try again.
+        if on_data_error == "warn":
+            message = f"Error processing tile, ENTIRE TILE LOST!!!: {ex}"
+            warnings.warn(message, UserWarning, stacklevel=3)
+
+            # Return
+            return_info["perfinfo"] = perfinfo
+            return_info["message"] = message
+            return return_info
+        else:
+            raise ex
+
     perfinfo["time_dissolve"] = (datetime.now() - start_dissolve).total_seconds()
 
     # If explodecollections is True and For polygons, explode multi-geometries.
@@ -1989,8 +2054,7 @@ def _dissolve(
     observed=False,
     dropna=True,
 ) -> gpd.GeoDataFrame:
-    """
-    Dissolve geometries within `groupby` into single observation.
+    """Dissolve geometries within `groupby` into single observation.
 
     This is accomplished by applying the `unary_union` method
     to all geometries within a groupself.
@@ -2081,9 +2145,9 @@ def _dissolve(
     data = pd.DataFrame(df.drop(columns=df.geometry.name))
 
     if aggfunc is not None and isinstance(aggfunc, dict) and "to_json" in aggfunc:
-        agg_columns = list(set(aggfunc["to_json"]))
+        agg_columns = list(aggfunc["to_json"])
         agg_data = (
-            data.groupby(**groupby_kwargs)
+            data.groupby(**groupby_kwargs)[agg_columns]
             .apply(lambda g: g[agg_columns].to_json(orient="records"))
             .to_frame(name="__DISSOLVE_TOJSON")
         )
@@ -2112,9 +2176,12 @@ def _dissolve(
             # Return as json string
             return json.dumps(json_distinct)
 
+        # Starting from pandas 2.2, include_groups=False should be passed to avoid
+        # warnings
+        kwargs = {"include_groups": False} if PANDAS_GTE_22 else {}
         agg_data = (
             data.groupby(**groupby_kwargs)
-            .apply(lambda g: group_flatten_json_list(g))
+            .apply(lambda g: group_flatten_json_list(g), **kwargs)
             .to_frame(name="__DISSOLVE_TOJSON")
         )
     else:
@@ -2132,7 +2199,11 @@ def _dissolve(
 
     # Process spatial component
     def merge_geometries(block):
-        merged_geom = block.unary_union
+        if GEOPANDAS_GTE_10:
+            merged_geom = block.union_all()
+        else:
+            merged_geom = block.unary_union
+
         return merged_geom
 
     g = df.groupby(group_keys=False, **groupby_kwargs)[df.geometry.name].agg(
