@@ -1,30 +1,28 @@
-"""
-Module containing utilities regarding sqlite/spatialite files.
-"""
+"""Module containing utilities regarding sqlite/spatialite files."""
 
 import datetime
 import enum
 import logging
-from pathlib import Path
 import pprint
 import shutil
 import sqlite3
 import tempfile
-from typing import Dict, List, Optional, Union
+from pathlib import Path
+from typing import Optional, Union
+
+from pygeoops import GeometryType
 
 import geofileops as gfo
-from geofileops import GeometryType
 from geofileops.helpers._configoptions_helper import ConfigOptions
-from geofileops.util._general_util import MissingRuntimeDependencyError
 from geofileops.util import _sqlite_userdefined as sqlite_userdefined
+from geofileops.util._general_util import MissingRuntimeDependencyError
 
 # Get a logger...
 logger = logging.getLogger(__name__)
 
 
 class EmptyResultError(Exception):
-    """
-    Exception raised when the SQL statement disn't return any rows.
+    """Exception raised when the SQL statement disn't return any rows.
 
     Attributes:
         message (str): Exception message
@@ -35,9 +33,8 @@ class EmptyResultError(Exception):
         super().__init__(self.message)
 
 
-def spatialite_version_info() -> Dict[str, str]:
-    """
-    Returns the versions of the spatialite modules.
+def spatialite_version_info() -> dict[str, str]:
+    """Returns the versions of the spatialite modules.
 
     Versions returned: spatialite_version, geos_version.
 
@@ -133,7 +130,7 @@ def get_columns(
     empty_output_ok: bool = True,
     use_spatialite: bool = True,
     output_geometrytype: Optional[GeometryType] = None,
-) -> Dict[str, str]:
+) -> dict[str, str]:
     # Create temp output db to be sure the output DB is writable, even though we only
     # create a temporary table.
     tmp_dir = Path(tempfile.mkdtemp(prefix="geofileops/get_columns_"))
@@ -292,8 +289,7 @@ def create_table_as_sql(
     column_datatypes: Optional[dict] = None,
     profile: SqliteProfile = SqliteProfile.DEFAULT,
 ):
-    """
-    Execute sql statement and save the result in the output file.
+    """Execute sql statement and save the result in the output file.
 
     Args:
         input1_path (Path): the path to the 1st input file.
@@ -596,7 +592,7 @@ def create_table_as_sql(
 
 
 def execute_sql(
-    path: Path, sql_stmt: Union[str, List[str]], use_spatialite: bool = True
+    path: Path, sql_stmt: Union[str, list[str]], use_spatialite: bool = True
 ):
     # Connect to database file
     conn = sqlite3.connect(path)
@@ -672,8 +668,7 @@ def test_data_integrity(path: Path, use_spatialite: bool = True):
 
 
 def load_spatialite(conn):
-    """
-    Load mod_spatialite for an existing sqlite connection.
+    """Load mod_spatialite for an existing sqlite connection.
 
     Args:
         conn ([type]): Sqlite connection
