@@ -1,12 +1,15 @@
+"""Run the benchmarks specified."""
+
 from benchmark import benchmarker
 
 
 def main():
-    # Run the benchmark function(s)
+    """Run the benchmark function(s)."""
     functions_to_run = [
-        "export_by_location_intersects",
-        #"join_by_location_intersects",
-        #"clip",
+        "export_by_location_intersects_complexpoly",
+        # "export_by_location_intersects",
+        # "join_by_location_intersects",
+        # "clip",
         # "intersection",
         # "intersection_complexpoly_agri",
         # "intersection_gridsize",
