@@ -1,3 +1,1 @@
-"""
-Sub-package containing helper modules that are quite specific for geofileops.
-"""
+"""Sub-package containing helper modules that are quite specific for geofileops."""
